More work [skip hazardlib]

+++ openquake/server/manage.py
@@ -27,6 +27,7 @@ except ImportError:
     def setproctitle(title):
         "Do nothing"
 from django.core.management import execute_from_command_line
+from openquake.baselib import config
 from openquake.server import dbserver
 from openquake.server.db import actions
 from openquake.commonlib import logs
@@ -63,9 +64,13 @@ if __name__ == "__main__":
     try:
         execute_from_command_line(sys.argv)
     finally:
-        for p in psutil.process_iter():
-            pname = p.name()
-            if pname == 'oq-worker':
-                os.kill(p.pid, signal.SIGKILL)
-            elif pname.startswith('oq-job-'):
-                os.kill(p.pid, signal.SIGKILL)
+        # do not kill anything in multiuser mode
+        if not config.dbserver.multi_user:
+            # kill all oq- processes spawned by the WebUI user
+            # SIGTERM is not enough and would leave defunct processes
+            for p in psutil.process_iter():
+                pname = p.name()
+                if pname == 'oq-worker':
+                    os.kill(p.pid, signal.SIGKILL)
+                elif pname.startswith('oq-job-'):
+                    os.kill(p.pid, signal.SIGKILL)
